step3_advanced_rag/requirements.txt
datasets==2.19.1
sentence-transformers==3.1.1
langchain==0.2.10
langchain-core==0.2.31
langchain-openai==0.2.2
faiss-cpu==1.8.0.post1
numpy==1.26.3
tqdm==4.66.4
colorama==0.4.7
